backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-ollama
ollama
python-multipart
pdfplumber
sentence-transformers
chromadb
numpy
fpdf2
